+++ apps/web/app/pdf/[...pdfId]/page.tsx
@@ -3,7 +3,7 @@ import { NotionAPI } from "notion-client";
 import { LessonView } from "@repo/ui/components";
 import { redirect } from "next/navigation";
 import { Print } from "../../../components/Print";
-import { getProblem, getTrack } from "../../tracks/[...trackIds]/page";
+import { getProblem, getTrack } from "../../../components/utils";
 
 const notion = new NotionAPI();
