+++ apps/web/components/utils.tsx
@@ -1,0 +1,45 @@
+import db from "@repo/db/client";
+
+export async function getProblem(problemId: string | null) {
+  if (!problemId) {
+    return null;
+  }
+  try {
+    const problem = await db.problem.findUnique({
+      where: {
+        id: problemId,
+      },
+    });
+    return problem;
+  } catch (err) {
+    return null;
+  }
+}
+
+export async function getTrack(trackId: string) {
+  try {
+    const track = await db.track.findUnique({
+      where: {
+        id: trackId,
+      },
+      include: {
+        problems: {
+          select: {
+            problem: true,
+          },
+        },
+      },
+    });
+
+    if (track) {
+      return {
+        ...track,
+        problems: track.problems.map((problem) => ({ ...problem.problem })),
+      };
+    }
+
+    return null;
+  } catch (err) {
+    return null;
+  }
+}
