+++ apps/web/tsconfig.json
@@ -6,6 +6,7 @@
         "name": "next",
       },
     ],
+    "strictNullChecks": true,
   },
   "include": [
     "next-env.d.ts",
