+++ turbo.json
@@ -1,10 +1,12 @@
 {
   "$schema": "https://turbo.build/schema.json",
-  "globalDependencies": ["**/.env.*local"],
+  "globalEnv": ["My_SECRET_TOKEN"],
+  "globalDependencies": ["**/.env.*local", ".env", ".env.local", "tsconfig.json"],
   "pipeline": {
     "build": {
       "dependsOn": ["^build"],
-      "outputs": [".next/**", "!.next/cache/**"]
+      "outputs": [".next/**", "!.next/cache/**"],
+      "env": ["My_SECRET_TOKEN"]
     },
     "lint": {
       "dependsOn": ["^lint"]
